Move CMakeConfiguration to own file.
